requirements.txt
Django==6.0.1
djangorestframework
djangorestframework-simplejwt
python-dotenv
requests
scikit-learn
numpy
psycopg2-binary
drf-spectacular
